Upgrade SDR pipeline from i8 to native i16 precision
The CPU pipeline now receives int16 samples from all backends,
preserving native ADC resolution:

- USRP: SC16 wire format (12-bit, full precision)
- bladeRF normal: SC16_Q11 shifted to fill i16 range (12-bit)
- bladeRF oversample: SC8 sign-extended (8-bit, no change)
- HackRF: CS8 sign-extended (8-bit, no change)
- SoapySDR: dynamic left-shift based on native format
- Spectran V6: f32 scaled to i16 (16-bit effective)
- GPU path unchanged (still i8 for OpenCL compatibility)

Recv batching accumulates ~4080 samples per channel send to
compensate for smaller per-recv batch sizes with wider formats.

Also adds receiver clock fallback for the Spectran V6 backend:
unsupported clocks are detected and the user is guided to valid
-C values for their device. Reference level is now tunable via -g.

Tested: USRP B210 (-C 40, -C 60), bladeRF 2.0 (-C 40),
Spectran V6 (-C 92), IQ file playback (ci8, ci16).

+++ crates/gpu/src/lib.rs
@@ -20,7 +20,7 @@ mod ffi {
             batch_size: c_uint,
         ) -> c_int;
 
-        pub fn gpu_pfb_get_buffer() -> *mut i8;
+        pub fn gpu_pfb_get_buffer() -> *mut i16;
         pub fn gpu_pfb_buffer_len() -> c_uint;
         pub fn gpu_pfb_submit() -> *mut c_float;
         pub fn gpu_pfb_flush() -> *mut c_float;
@@ -106,12 +106,12 @@ impl GpuChannelizer {
         })
     }
 
-    /// Get a mutable slice to the raw int8 buffer for the current batch.
+    /// Get a mutable slice to the raw int16 buffer for the current batch.
     ///
-    /// Fill this with `batch_size * M` int8 IQ bytes (M/2 complex samples per step,
-    /// 2 bytes per complex sample, M bytes per step). The pre-roll region is already
-    /// filled from the previous batch.
-    pub fn raw_buffer(&self) -> &mut [i8] {
+    /// Fill this with `batch_size * M` int16 IQ samples (M/2 complex samples per step,
+    /// 2 int16 values per complex sample, M int16 values per step). The pre-roll region
+    /// is already filled from the previous batch.
+    pub fn raw_buffer(&self) -> &mut [i16] {
         unsafe {
             let ptr = ffi::gpu_pfb_get_buffer();
             slice::from_raw_parts_mut(ptr, self.buffer_len)
